change to integer

+++ Task06/src/main/java/by/golik/jwdcourse/task06/query/search_query/SearchPagesQuery.java
@@ -18,13 +18,13 @@ public class SearchPagesQuery implements BookQuery {
         return book.getPages() == pages;
     }
 
-    private final Long pages;
+    private final int pages;
 
     /**
      * Constructor
      * @param identify - count of pages
      */
-    public SearchPagesQuery(final Long identify) {
+    public SearchPagesQuery(final int identify) {
         this.pages = identify;
     }
